requirements.txt
Django>=1.8,<2.0
django-mutant==0.2.1
psycopg2==2.7.3
